move class definition into header.
