+++ src/search/open_lists/tiebreaking_open_list.cc
@@ -15,44 +15,6 @@
 using namespace std;
 
 namespace tiebreaking_open_list {
-template<class Entry>
-class TieBreakingOpenList : public OpenList<Entry> {
-    using Bucket = deque<Entry>;
-
-    map<const vector<int>, Bucket> buckets;
-    int size;
-
-    vector<shared_ptr<Evaluator>> evaluators;
-    /*
-      If allow_unsafe_pruning is true, we ignore (don't insert) states
-      which the first evaluator considers a dead end, even if it is
-      not a safe heuristic.
-    */
-    bool allow_unsafe_pruning;
-
-    int dimension() const;
-
-protected:
-    virtual void do_insertion(EvaluationContext &eval_context,
-                              const Entry &entry) override;
-
-public:
-    explicit TieBreakingOpenList(const plugins::Options &opts);
-    explicit TieBreakingOpenList(
-        bool pref_only,
-        vector<shared_ptr<Evaluator>> evaluators,
-        bool allow_unsafe_pruning);
-    virtual ~TieBreakingOpenList() override = default;
-
-    virtual Entry remove_min() override;
-    virtual bool empty() const override;
-    virtual void clear() override;
-    virtual void get_path_dependent_evaluators(set<Evaluator *> &evals) override;
-    virtual bool is_dead_end(
-        EvaluationContext &eval_context) const override;
-    virtual bool is_reliable_dead_end(
-        EvaluationContext &eval_context) const override;
-};
 
 
 template<class Entry>
